backend/foodgram/requirements.txt
Django==3.2
python-dotenv
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg-binary==3.0.10
django-colorfield==0.6.3
psycopg2-binary==2.9.3
djoser==2.1.0
webcolors==1.11.1
drf-extra-fields==3.4.0
django-filter==21.1
